Refacted view models into modules
